backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
pydantic==2.10.3
PyJWT==2.10.1
psycopg[binary]==3.2.3
